refactor: moved RasterTileMetadata related inteface to svelte-undp-components
