+++ packages/svelte-undp-components/src/lib/interfaces/index.ts
@@ -2,3 +2,4 @@ export * from './VectorLayerMetadata.js';
 export * from './VectorLayerTileStatAttribute.js';
 export * from './VectorLayerTileStatLayer.js';
 export * from './VectorTileMetadata.js';
+export * from './RasterTileMetadata.js';
