+++ sites/geohub/src/components/pages/map/data/DataCard.svelte
@@ -10,13 +10,14 @@
 	import { VectorTileData } from '$lib/VectorTileData';
 	import { AccessLevel } from '$lib/config/AppConfig';
 	import { getFirstSymbolLayerId, isRgbRaster, loadMap } from '$lib/helper';
-	import type { DatasetFeature, Layer, LayerCreationInfo, RasterTileMetadata } from '$lib/types';
+	import type { DatasetFeature, Layer, LayerCreationInfo } from '$lib/types';
 	import { LAYERLISTSTORE_CONTEXT_KEY, type LayerListStore } from '$stores';
 	import {
 		Accordion,
 		initTooltipTippy,
 		MAPSTORE_CONTEXT_KEY,
 		type MapStore,
+		type RasterTileMetadata,
 		type VectorLayerTileStatLayer,
 		type VectorTileMetadata
 	} from '@undp-data/svelte-undp-components';
