+++ sites/geohub/src/components/pages/map/data/DataVectorCard.svelte
@@ -7,12 +7,13 @@
 	import { VectorTileData } from '$lib/VectorTileData';
 	import type { UserConfig } from '$lib/config/DefaultUserConfig';
 	import { loadMap } from '$lib/helper';
-	import type { DatasetFeature, LayerCreationInfo, RasterTileMetadata } from '$lib/types';
+	import type { DatasetFeature, LayerCreationInfo } from '$lib/types';
 	import { LAYERLISTSTORE_CONTEXT_KEY, type LayerListStore } from '$stores';
 	import {
 		Accordion,
 		MAPSTORE_CONTEXT_KEY,
 		type MapStore,
+		type RasterTileMetadata,
 		type VectorLayerTileStatLayer,
 		type VectorTileMetadata
 	} from '@undp-data/svelte-undp-components';
