+++ sites/geohub/src/components/pages/map/layers/LayerInfo.svelte
@@ -1,12 +1,13 @@
 <script lang="ts">
 	import Star from '$components/util/Star.svelte';
 	import { createAttributionFromTags, isRgbRaster } from '$lib/helper';
-	import type { Layer, RasterTileMetadata } from '$lib/types';
+	import type { Layer } from '$lib/types';
 	import {
 		Accordion,
 		FieldControl,
 		Histogram,
-		initTooltipTippy
+		initTooltipTippy,
+		type RasterTileMetadata
 	} from '@undp-data/svelte-undp-components';
 	import { CtaLink } from '@undp-data/svelte-undp-design';
 	import { filesize } from 'filesize';
