+++ sites/geohub/src/components/pages/map/plugins/MapQueryInfoControl.svelte
@@ -5,14 +5,16 @@
 		getLayerStyle,
 		getValueFromRasterTileUrl
 	} from '$lib/helper';
-	import type { BandMetadata, Layer, RasterTileMetadata } from '$lib/types';
+	import type { Layer } from '$lib/types';
 	import type { LayerListStore } from '$stores';
 	import {
 		Accordion,
 		clean,
 		handleEnterKey,
 		initTooltipTippy,
-		isValidUrl
+		isValidUrl,
+		type BandMetadata,
+		type RasterTileMetadata
 	} from '@undp-data/svelte-undp-components';
 	import { Checkbox, DefaultLink, Loader } from '@undp-data/svelte-undp-design';
 	import { Map, MapMouseEvent, Popup, type ControlPosition, type PointLike } from 'maplibre-gl';
