+++ sites/geohub/src/components/util/stac/StacApiExplorer.svelte
@@ -17,7 +17,6 @@
 		Layer,
 		LayerCreationInfo,
 		RasterAlgorithm,
-		RasterTileMetadata,
 		Stac,
 		StacItemFeatureCollection,
 		StacProduct,
@@ -28,6 +27,7 @@
 		DatePicker,
 		FieldControl,
 		Notification,
+		type RasterTileMetadata,
 		SegmentButtons,
 		ShowDetails,
 		Slider,
