+++ sites/geohub/src/lib/server/defaultStyle/DefaultStyleTemplate.ts
@@ -1,6 +1,6 @@
 import type { UserConfig } from '$lib/config/DefaultUserConfig';
-import type { DatasetDefaultLayerStyle, DatasetFeature, RasterTileMetadata } from '$lib/types';
-import type { VectorTileMetadata } from '@undp-data/svelte-undp-components';
+import type { DatasetDefaultLayerStyle, DatasetFeature } from '$lib/types';
+import type { VectorTileMetadata, RasterTileMetadata } from '@undp-data/svelte-undp-components';
 
 export interface DefaultStyleTemplate {
 	dataset: DatasetFeature;
