+++ sites/geohub/src/lib/server/helpers/getRasterMetadata.ts
@@ -1,8 +1,11 @@
-import type { BandMetadata, RasterTileMetadata } from '$lib/types';
 import { generateAzureBlobSasToken } from '$lib/server/helpers';
 import { env } from '$env/dynamic/private';
 import { getBase64EncodedUrl } from '$lib/helper';
-import { clean } from '@undp-data/svelte-undp-components';
+import {
+	clean,
+	type BandMetadata,
+	type RasterTileMetadata
+} from '@undp-data/svelte-undp-components';
 
 export const getRasterMetadata = async (url: string) => {
 	const sasToken = await generateAzureBlobSasToken(url);
