+++ sites/geohub/src/lib/types/LayerCreationInfo.ts
@@ -4,10 +4,10 @@ import type {
 	VectorSourceSpecification
 } from 'maplibre-gl';
 import type { VectorLayerSpecification } from './VectorLayerSpecification';
-import type { RasterTileMetadata } from './RasterTileMetadata';
 import type {
 	VectorTileMetadata,
-	ClassificationMethodTypes
+	ClassificationMethodTypes,
+	RasterTileMetadata
 } from '@undp-data/svelte-undp-components';
 import type { DatasetDefaultLayerStyle } from '.';
