--- sites/geohub/src/lib/types/RasterLayerStats.ts
@@ -1,5 +1,0 @@
-import type { BandStatistics } from './BandStatistics';
-
-export interface RasterLayerStats {
-	[band_no: string]: BandStatistics;
-}
