+++ sites/geohub/src/lib/types/index.ts
@@ -1,8 +1,6 @@
 export * from './AcceptedExtension';
 export * from './ArithmeticOperator';
 export * from './AssetOptions';
-export * from './BandMetadata';
-export * from './BandStatistics';
 export * from './Continent';
 export * from './Country';
 export * from './DashboardMapStyle';
@@ -23,9 +21,7 @@ export * from './PgtileservDetailJson';
 export * from './PgtileservIndexJson';
 export * from './Product';
 export * from './RasterAlgorithm';
-export * from './RasterLayerStats';
 export * from './RasterSimpleExpression';
-export * from './RasterTileMetadata';
 export * from './Region';
 export * from './Sprite';
 export * from './SpriteIcon';
